Create checkrbl.sh
CheckRBL.sh simply checks the logs to see how effective the RBL's are at blocking inbound mail. It also checks for mails rejected for failing a reverse dns check. This is inspired by PFLogSumm which is a great perl script for doing the exact same thing way more verbosely and thoroughly, but this is in an easier to read format.